[pipes] golden test suite: initial setup
