+++ libraries/pipes/implementations/java/dagster-pipes-java/pyproject.toml
@@ -21,7 +21,11 @@ dev-dependencies = [
     "moto[server,s3]>=5.0.21",
     "boto3[s3]>=1.35.67",
     "boto3-stubs[s3]>=1.35.67",
+    "dagster-pipes-tests",
 ]
+
+[tool.uv.sources]
+dagster-pipes-tests = { path = "../../../tests/dagster-pipes-tests" }
 
 [tool.pytest.ini_options]
 addopts = "-s "
