+++ libraries/pipes/tests/dagster-pipes-tests/README.md
@@ -1,0 +1,11 @@
+# Dagster Pipes Tests
+
+This is a "golden" test suite for the Dagster Pipes protocol. It is designed to test the protocol implementation in various languages, ensuring that the protocol is implemented correctly and consistently across all them.
+
+# Usage
+
+To make use of this test suite, it has to be installed as a Python package.
+
+The package is a `pytest` plugin which provides various fixtures and utilities for testing Pipes.
+
+TODO: Add more information on how to use the test suite.
