+++ libraries/pipes/tests/dagster-pipes-tests/pyproject.toml
@@ -1,0 +1,22 @@
+[project]
+name = "dagster-pipes-tests"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+authors = [
+    { name = "danielgafni", email = "danielgafni16@gmail.com" }
+]
+requires-python = ">=3.12"
+dependencies = [
+    "pytest>=8.3.4",
+]
+
+[project.scripts]
+dagster-pipes-tests = "dagster_pipes_tests:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project.entry-points.pytest11]
+dagster-pipes-tests = "dagster_pipes_tests"
