+++ libraries/pipes/tests/dagster-pipes-tests/src/dagster_pipes_tests/cases/context_injector.py
